requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
streamlit==1.39.0
pymupdf==1.24.10
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-multipart==0.0.9